+++ PiecesOfPaperTests/NoteRepositoryMock.swift
@@ -188,82 +188,3 @@ final class NoteRepositoryMock: NoteRepositoryProtocol {
         completion(NoteData(entity: NoteEntity(drawing: note.entity.drawing), fileURL: newUrl))
     }
 }
-
-/// Stands in for the on-disk cache file: save() replaces what load() returns,
-/// so a store can be started "warm" by seeding entries.
-final class NoteMetadataCacheRepositoryMock: NoteMetadataCacheRepositoryProtocol, @unchecked Sendable {
-    private let lock = NSLock()
-    private var storage: [String: NoteMetadata]
-    private var saves = 0
-
-    init(entries: [String: NoteMetadata] = [:]) {
-        storage = entries
-    }
-
-    var savedEntries: [String: NoteMetadata] { lock.withLock { storage } }
-    var saveCount: Int { lock.withLock { saves } }
-
-    func load() -> [String: NoteMetadata] { lock.withLock { storage } }
-
-    func save(_ entries: [String: NoteMetadata]) {
-        lock.withLock {
-            storage = entries
-            saves += 1
-        }
-    }
-}
-
-struct UbiquityStatusMock: UbiquityStatusProviding {
-    var hasAccount = true
-    var containerUrl: URL? = URL(fileURLWithPath: "/container/Documents")
-}
-
-final class PreferenceRepositoryMock: PreferenceRepositoryProtocol {
-    var enablediCloud = false
-    var enabledAutoSave = true
-    var enabledInfiniteScroll = true
-    var appearanceMode: AppearanceMode = .system
-    var listOrders: [String: ListOrder] = [:]
-    private(set) var setEnablediCloudCalls: [Bool] = []
-    private(set) var setEnabledAutoSaveCalls: [Bool] = []
-    private(set) var setEnabledInfiniteScrollCalls: [Bool] = []
-    private(set) var setAppearanceModeCalls: [AppearanceMode] = []
-    private(set) var setListOrderCalls: [(directoryName: String, listOrder: ListOrder)] = []
-
-    func getEnablediCloud() -> Bool { enablediCloud }
-
-    func setEnablediCloud(_ value: Bool) {
-        enablediCloud = value
-        setEnablediCloudCalls.append(value)
-    }
-
-    func getEnabledAutoSave() -> Bool { enabledAutoSave }
-
-    func setEnabledAutoSave(_ value: Bool) {
-        enabledAutoSave = value
-        setEnabledAutoSaveCalls.append(value)
-    }
-
-    func getEnabledInfiniteScroll() -> Bool { enabledInfiniteScroll }
-
-    func setEnabledInfiniteScroll(_ value: Bool) {
-        enabledInfiniteScroll = value
-        setEnabledInfiniteScrollCalls.append(value)
-    }
-
-    func getAppearanceMode() -> AppearanceMode { appearanceMode }
-
-    func setAppearanceMode(_ value: AppearanceMode) {
-        appearanceMode = value
-        setAppearanceModeCalls.append(value)
-    }
-
-    func getListOrder(directoryName: String) -> ListOrder {
-        listOrders[directoryName] ?? ListOrder()
-    }
-
-    func setListOrder(directoryName: String, listOrder: ListOrder) {
-        listOrders[directoryName] = listOrder
-        setListOrderCalls.append((directoryName, listOrder))
-    }
-}
